requirements.txt
amadeus==5.1.0
backoff==1.10.0
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
opencage==1.2.2
psycopg2==2.9.1
pycparser==2.20
pyOpenSSL==20.0.1
python-restcountries==1.0.0
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.20
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
